+++ unordered_arrays.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "unordered_arrays", "unordered_arrays\unordered_arrays.vcxproj", "{BF613583-8B63-4D5C-8645-244601BD5B96}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BF613583-8B63-4D5C-8645-244601BD5B96}.Debug|x64.ActiveCfg = Debug|x64
+		{BF613583-8B63-4D5C-8645-244601BD5B96}.Debug|x64.Build.0 = Debug|x64
+		{BF613583-8B63-4D5C-8645-244601BD5B96}.Debug|x86.ActiveCfg = Debug|Win32
+		{BF613583-8B63-4D5C-8645-244601BD5B96}.Debug|x86.Build.0 = Debug|Win32
+		{BF613583-8B63-4D5C-8645-244601BD5B96}.Release|x64.ActiveCfg = Release|x64
+		{BF613583-8B63-4D5C-8645-244601BD5B96}.Release|x64.Build.0 = Release|x64
+		{BF613583-8B63-4D5C-8645-244601BD5B96}.Release|x86.ActiveCfg = Release|Win32
+		{BF613583-8B63-4D5C-8645-244601BD5B96}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
